💄 style: /mypage UI 깨지는 부분들 수정
