+++ app/_components/Tabs.tsx
@@ -3,11 +3,6 @@ import { ReactNode } from "react";
 import TabButton from "@/components/TabButton";
 import { Tab } from "@/store/tab/cookies";
 
-const TOP_OFFSET = {
-  mypage: "top-72 tablet:top-64",
-  event: "top-60",
-};
-
 interface Props {
   children: ReactNode[];
   names: readonly string[];
@@ -20,7 +15,7 @@ const Tabs = ({ children, names, page = "mypage", eventId, isNarrow: narrow = fa
 
   return (
     <section className="h-full w-full">
-      <div className={`sticky top-72 z-popup flex h-44 w-full items-center border-b border-gray-50 bg-white-black px-20 pc:top-64 pc:h-56 ${TOP_OFFSET} ${narrow && "pc:static"}`}>
+      <div className={`sticky top-72 z-popup flex h-44 w-full items-center border-b border-gray-50 bg-white-black px-20 tablet:top-64 pc:h-56 ${narrow && "pc:static"}`}>
         {names.map((name, index) => (
           <TabButton key={name} tab={{ name, index }} selected={name === tab.name} narrow={narrow}>
             {name}
